chore(search): Implement search service
