+++ src/pages/search-chat.tsx
@@ -2,9 +2,18 @@ import { useState } from "react";
 import { Command, CommandDialog, CommandInput, CommandList, CommandItem, CommandEmpty } from "@/components/ui/command";
 import { Input } from "@/components/ui/input";
 import { useNavigate } from "react-router-dom";
-import { type IUser, useAuth } from "@/context/AuthContext";
-import { collection, query, where, getDocs } from 'firebase/firestore';
-import { db } from '../config/firebase';
+import { useAuth } from "@/context/AuthContext";
+import { IUser } from "@/interfaces/auth";
+import { searchService } from "@/server/api/search";
+
+const debounce = <T extends (...args: any[]) => void>(cb: T, wait: number = 500): ((...args: Parameters<T>) => void) => {
+  let timeout: ReturnType<typeof setTimeout>;
+
+  return (...args: Parameters<T>) => {
+    clearTimeout(timeout);
+    timeout = setTimeout(() => cb(...args), wait);
+  };
+};
 
 const SearchChatPage = () => {
   const { user } = useAuth();
@@ -13,60 +22,14 @@ const SearchChatPage = () => {
   const [open, setOpen] = useState(false);
   const [results, setResults] = useState<IUser[]>([]);
 
-  const debounce = <T extends (...args: any[]) => void>(cb: T, wait: number = 500): ((...args: Parameters<T>) => void) => {
-    let timeout: ReturnType<typeof setTimeout>;
-
-    return (...args: Parameters<T>) => {
-      clearTimeout(timeout);
-      timeout = setTimeout(() => cb(...args), wait);
-    };
-  };
-
   const searchUsers = async (searchTerm: string) => {
-    if (!searchTerm) {
-      return [];
-    };
-
-    const usersRef = collection(db, 'users');
-
-    // Single field search to avoid composite index requirement
-    const emailQuery = query(
-      usersRef,
-      where('email', '>=', searchTerm.toLowerCase()),
-      where('email', '<=', searchTerm.toLowerCase() + '\uf8ff')
-    );
-
-
-    const nameQuery = query(
-      usersRef,
-      where('displayName', '>=', searchTerm),
-      where('displayName', '<=', searchTerm + '\uf8ff')
-    );
-
-    const [emailSnapshot, nameSnapshot] = await Promise.all([
-      getDocs(emailQuery),
-      getDocs(nameQuery)
-    ]);
-
-
-    // Combine and deduplicate results
-    const results = new Map();
-
-    emailSnapshot.forEach((doc) => {
-      results.set(doc.id, { id: doc.id, ...doc.data() });
-    });
-
-    nameSnapshot.forEach((doc) => {
-      results.set(doc.id, { id: doc.id, ...doc.data() });
-    });
-
-    return Array.from(results.values());
+    if (!searchTerm) return [];
+    return searchService.searchUsers(searchTerm);
   };
 
   const handleValueChange = debounce(async (value: string) => {
     const results = await searchUsers(value);
     setResults(results);
-
   });
 
   return (
